+++ src/base/TgaReader.cpp
@@ -276,7 +276,7 @@ static void ReadPixel(ReadState& s, u8* dst, int bits, int alphaBits, ImageAlpha
         case Type_Palette_RLE:
             idx = ((u8)s.data[0] | (2 == s.n ? ((u8)s.data[1] << 8) : 0)) - s.cmap.firstEntry;
             if (0 <= idx && idx < s.cmap.length) {
-                src = s.cmap.data + (idx * s.cmap.n);
+                src = s.cmap.data + ((size_t)idx * s.cmap.n);
             } else {
                 s.failed = true;
             }
@@ -320,7 +320,7 @@ Pixmap* PixmapFromData(Str d) {
         s.cmap.n = (headerLE->cmapBitDepth + 7) / 8;
         s.cmap.length = convLE(headerLE->cmapLength);
         s.cmap.firstEntry = convLE(headerLE->cmapFirstEntry);
-        s.data += s.cmap.length * s.cmap.n;
+        s.data += (size_t)s.cmap.length * s.cmap.n;
         if (s.data > s.end) {
             return nullptr;
         }
@@ -349,9 +349,9 @@ Pixmap* PixmapFromData(Str d) {
         return nullptr;
     }
     for (int y = 0; y < h; y++) {
-        u8* rowOut = pixmap->data + (pixmap->stride * (invertY ? y : h - 1 - y));
+        u8* rowOut = pixmap->data + ((size_t)pixmap->stride * (invertY ? y : h - 1 - y));
         for (int x = 0; x < w; x++) {
-            ReadPixel(s, rowOut + (4 * (invertX ? w - 1 - x : x)), bits, alphaBits, alphaType);
+            ReadPixel(s, rowOut + ((size_t)4 * (invertX ? w - 1 - x : x)), bits, alphaBits, alphaType);
         }
     }
     if (s.failed) {
